Piece 타입 별 클래스 코드 분리
