+++ ChessApp/ChessApp/Bishop.swift
@@ -1,0 +1,26 @@
+//
+//  Bishop.swift
+//  ChessApp
+//
+//  Created by elly on 11/1/23.
+//
+
+import Foundation
+
+class Bishop: Piece {
+    var pieceType: PieceType = .bishop
+    var point: Int = 3
+    var pieceColorType: PieceColorType
+    
+    init(pieceColorType: PieceColorType) {
+        self.pieceColorType = pieceColorType
+    }
+    
+    func showIcon() -> String {
+        return pieceColorType == .black ? "♝" : "♗"
+    }
+    
+    func movablePosition(_ currentPosition: Position) -> [Position] {
+        return currentPosition.diagonal()
+    }
+}
