+++ ChessApp/ChessApp/PieceType.swift
@@ -1,0 +1,15 @@
+//
+//  PieceType.swift
+//  ChessApp
+//
+//  Created by elly on 11/1/23.
+//
+
+import Foundation
+
+enum PieceType {
+    case empty
+    case pawn
+    case bishop
+    case rook
+}
